Move visualizations to Jupyter project
