+++ src/Jupyter/Visualization/DataTableEncoders.cs
