+++ src/Jupyter/Visualization/DisplayableExceptionEncoders.cs
