+++ src/Jupyter/Visualization/TaskStatusEncoders.cs
@@ -6,7 +6,7 @@
 using System;
 using Microsoft.Jupyter.Core;
 
-namespace Microsoft.Quantum.IQSharp.Kernel
+namespace Microsoft.Quantum.IQSharp.Jupyter
 {
 
     /// <summary>
