+++ src/Kernel/Magic/PackageMagic.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Reactive.Linq;
 using System.Threading.Tasks;
 using Microsoft.Jupyter.Core;
+using Microsoft.Quantum.IQSharp.Jupyter;
 using Microsoft.Quantum.IQSharp.Kernel;
 
 namespace Microsoft.Quantum.IQSharp.Kernel
@@ -39,7 +40,7 @@ namespace Microsoft.Quantum.IQSharp.Kernel
         public override ExecutionResult Run(string input, IChannel channel)
         {
             var (name, _) = ParseInput(input);
-            var status = new TaskStatus($"Adding package {name}");
+            var status = new Jupyter.TaskStatus($"Adding package {name}");
             var statusUpdater = channel.DisplayUpdatable(status);
             void Update() => statusUpdater.Update(status);
