+++ src/Tests/DisplayConverterTests.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Microsoft.Quantum.IQSharp.Kernel;
+using Microsoft.Quantum.IQSharp.Jupyter;
 using System.Linq;
 using System.Collections.Generic;
